requirements.txt
# Web Framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Configuration
python-dotenv>=1.0.1
pydantic>=2.9.0
pydantic[email]>=2.9.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0

# OpenAI & External APIs
openai>=1.30.0
httpx>=0.27.0
requests>=2.31.0

# PDF & Documents
reportlab>=4.0.0

# Email
aiosmtplib>=3.0.0

# Development (opcional pero recomendado)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

